Updated the previous code file named jsc to include algorithm for scheduling
Added an algorithm which is a modified first-fit.
It first checks the list for servers with available resources for job. Than it adds it to a tempList.
Than we check which servers in templList are active or idle.
If none of them are active or idle, we schedule it to the first element in tempList.